Add the Packet Capture Specification document
This document provides a specification of the format of the CSV
capture files we will be using as test data, as well as an index of
the collected data. The document should be updated with any new data
we collect as it provides a record of our test data for future
reference.