New Visualization Framework
-- Changed the visualization classes. Now they extend Visualization
directly (no interface).
-- If you have been working on a visualization just extend Visualization
instead of AbstractVisualization and do the initial draw in init()
instead of verifying if firstDraw is true;
-- Each visualization is now a GLCanvas! The base class makes sure that
redrawing is done only in the active visualization. The dummy
visualization is still in place. I've done this because each
visualization might want to do special stuff with the camera / jogl
painter and it's hard to reset everything when changing the active
visualization.
-- Casted the list of packets returned by the DataController into a
immutable list (should have done this from the beginning to protect that
list outside the class.)
-- Added a "cube" visualization. This visualization is still in
progress. However you can take a look and see how it's implemented for a
greater understaning of how a visualization works.

To be optimized:
-- Now that every visualization is a GLCanvas it is more inefficient to
instantiate all the visualizations when the app starts. A factory method
for the visualizations should be created and used in RightPanel.

Note:
-- If you are not satisfied with the new way we can roll back to the
last commit. However it will be much harder to implement different
visualizations using the old way since they are not really independent.