Make the context panel a global object
The context panel is now a final field in ApplicationFrame.
This means you can pass a reference to it to your visualisations, and call
contextPanel.update() as you see fit, e.g. when a user clicks on part of
your visualisation, you can pass a panel to give details of the selected
part.