Write User Workflow + Implement Suggestions
Where suggestions from DP haven't been carried out to the letter, I've included
a comment stating why not.

Updated spellings to 'Conformist British English' standard

I still disagree with the use of capitals for 'Traffic Volume visualization',
we haven't created a product with this name, so it's not a pronoun and
shouldn't be treated as such.

Anyhow, if we're going with this, the 'Dataflow' subheading needs to be changed
to be consistent with the text body.